Split constraints into submodule
